Modificarea logicii de creare a orarului. Adaugat prelucrare pentru <30, >30/<60, 60+

+++ Disciplina.cs
@@ -3,9 +3,9 @@
     public class Disciplina
     {
         public string Denumire { get; set; }
-        public int OreCurs { get; set; }
-        public int OreSeminar { get; set; }
-        public int OreLaborator { get; set; }
+        public double OreCurs { get; set; }
+        public double OreSeminar { get; set; }
+        public double OreLaborator { get; set; }
 
         public Disciplina(string nume, int oreCurs, int oreSeminar, int oreLaborator)
         {
